fix: split up code
